Move ISuppressedIssueMatcher, its implementation and its test into the ConnectedMode assembly
