+++ src/ConnectedMode.UnitTests/Suppressions/SuppressedIssueMatcherTests.cs
